Add GA feature selection
A competing approach: Wrapper feature selection with genetic algorithms,
as provided by the caret package. Code a bit ugly as we try to cut down
caret's excessive parameter tuning and evaluation to the necessary minimum.
We use MCC for evaluation and a penalized MCC as fitness criterion
(incentivize the selection of a small feature set). kNN and xgboost have
been tested as wrapped models.