Add simple penalized selection strategy
Instead of finding the best k of m features with a fixed k, run all subsets
with size 1..m and select the best subset according to a penalized
performance metric, discounting the number features. The progress bar is
not reliable for this approach, as the number of features selected per
micro-task can vary, it might even be that no feature is discarded (could
prevent convergence).