Add model-based selection strategy
Another strategy with variable k (besides the penalized one). In each
m-subset, all features are selected which are used by the wrapped model
(here: xgboost). As this strategy might simply select all features (as the
penalized strategy), we add an additional check on the number of selected
features in the main loop.